marketmuse/requirements.txt
crewai
pydantic
python-dotenv
crewai[google]
crewai-tools